[cli] move NativeFileSystemImpl and Simulator back to peergoes.tests.simulation
